% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biteTiming.R
\name{biteTiming}
\alias{biteTiming}
\title{biteTiming: Calculates bite timing from entered cumulative intake. If no cumulative intake data is entered, it will simulated a bite dataset.}
\usage{
biteTiming(
  timeDat,
  nBites,
  Emax,
  mealDure,
  timePDF,
  parameters,
  model_str = "LODE",
  id,
  procNoise = TRUE,
  pNoiseSD
)
}
\arguments{
\item{nBites}{A numeric value for total number of bites in a meal.}

\item{Emax}{A numeric value for total cumulative intake.}

\item{timePDF}{A string or vector of strings for the generating probibility distribution function. Options include: 'logis', 'quad', 'u-quad', 'exp', or 'linear'. Default is 'logis'. Can enter more than one to return different Time variables.}

\item{parameters}{A set of numeric parameters: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order (default is c(10, 1, -1)) and Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order (default is c(10, .10)).}

\item{model_str}{The base model to use--'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'.}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}

\item{procNoise}{(optional) A logical indicator for adding random process noise to the timing data by jittering bite timing with cumulative intake calculated from jittered bite timing. This uses the default jitter amount (smallest distance/5). Default value is TRUE if bite timings are simulated.}

\item{pNoiseSD}{(optional) This allows you to enter the standard deviation of individuals bites sizes and will replace the default procNoise routine (jittered bite sizes). Bite sizes will be randomly chosen from a normal distribution truncated at min = 0 with mean = Emax/nBites and standard deviation equal to the entered value. procNoise must be set to TRUE, otherwise this argument will be ignored.}

\item{intaketimeDat}{If a vector of cumulative intake is entered, will calculate the bite timing using entered parameters rather than simulating timing data. No process noise will be added.}
}
\value{
A bite dataset with bite timing, bite size, and cumulative intake for each bite
}
\description{
This function calcualtes bite timing from cumulative intake data using the cumulative intake
data and the specified parameters of the model. The intake data will be calculated using either
the Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001) or the Logistic Ordinary
Differential Equation (LODE) model (Thomas et al., 2017). If no bite timings are entered,
this function simulates a bite dataset from specified parameters and model.
}
\details{
The simulation randomly samples bite timing using an uniformed distribution.
There is the option of adding process noise to the bite timing, which can be added through
jitter or a specified standard deviation. If sd_bitesize is specified, the bites timings
will randomly vary across the meal using a Gaussian distribution with mean = timing between bites
and SD = pNoiseSD. Process noise is added *before* to the calculation of bite timing.
The timing of each bite will be calculated using the specified model and parameters
}
\examples{

\dontrun{
}


}
\references{
Fogel A, Goh AT, Fries LR, et al. Physiology & Behavior. 2017;176:107-116
(\href{https://pubmed.ncbi.nlm.nih.gov/28213204/}{PubMed})
Kissileff HR, Thorton J, Becker E. Appetite. 1982;3:255-272
(\href{https://pubmed.ncbi.nlm.nih.gov/7159076/}{PubMed})
Kissileff HR, Guss JL. Appetite. 2001;36:70-78
(\href{https://pubmed.ncbi.nlm.nih.gov/11270360/}{PubMed})
Thomas DM, Paynter J, Peterson CM, et al. Am J Clin Nutr. 2017;105:323-331
 (\href{https://pubmed.ncbi.nlm.nih.gov/28077377/}{PubMed})
}
