% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_LPE.R
\name{CI_LPE}
\alias{CI_LPE}
\title{CI_LPE: Calculates log-likelihood ratio test for the upper or lower confidence bound}
\usage{
CI_LPE(
  data,
  par,
  Emax,
  model_str = "LODE",
  min_n2ll,
  paramIndex,
  conf = 95,
  bound,
  timeVar,
  intakeVar
)
}
\arguments{
\item{data}{A data frame that contains:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{par}{A set of numeric parameters: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order; the Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order}

\item{Emax}{Total cumulative intake}

\item{model_str}{(optional) Only needed if simulating data. 'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'}

\item{min_n2ll}{The minimum -2 log-likelihood value obtained when fitting the parameters}

\item{paramIndex}{The index number for par that corresponds to the parameter the CI is being fit for. E.g., if LODE Model, par[1] would be theta and par[2] would be r}

\item{conf}{(optional) Level of confidence for calculation of confidence intervals around the fitted parameter estimates. Default is 95 for 95\% CI. If no confidence intervals are desired, set conf = NA}

\item{bound}{A string with the boundary value desired: 'upper' or 'lower'}

\item{timeVar}{(optional) Use if name of bite timing variable does not include 'time' or multiple variable names in data include the string 'time'}

\item{intakeVar}{(optional) Use if name of bite timing variable does not include 'cumulative' or multiple variable names in data include the string 'cumulative'}
}
\value{
The likelihood ratio test for the CI bound and value (upper v lower) requested
}
\description{
Calculates log-likelihood ratio test for the upper or lower confidence bound.
}
\examples{

\dontrun{
}

}
