% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_LRTest.R
\name{CI_LRTest}
\alias{CI_LRTest}
\title{CI_LRTest: Wrapper of model fit functions that calculates the likelihood ratio test}
\usage{
CI_LRTest(
  data,
  par,
  n2ll_fn = FPM_2nll,
  timeVar,
  intakeVar,
  min_n2ll,
  paramIndex,
  bound
)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{par}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic)}

\item{n2ll_fn}{Name of the function for the -2LL calculation for given model (i.e., FPM_n2ll or Kissileff_n2ll)}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{min_n2ll}{This is a single value or list of the fitted minimum -2 log-likelihood. This will be used as the -2 log-likelihood value for all fitted parameter values entered in paramCI. If multiple values are entered, must enter an equal number of parameter values in paramCI for each labeled parameter. E.g., paramCI = list(r = c(0.10, 0.15)) if enter two -2 log-likelihood values.}

\item{paramIndex}{The index number for par that corresponds to the parameter the CI is being fit for. E.g., if First Principles Model, par[1] would be theta and par[2] would be r.}

\item{bound}{A string indicating which confidence bound to return: 'upper, 'lower', or 'both'. Default = 'both'}
}
\value{
NEED TO EDIT
}
\description{
This function wrpas model fit functions called through optim {stats} to calcualte the likelihood ratio test. This is a function that, when called by optim {stats} will identify a confidence bound of a parameter by optim's minimization of the calculated LRT value.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Fit}}.
}
