% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fogel2017_rmvnMicroBeh.R
\name{Fogel2017_rmvnMicroBeh}
\alias{Fogel2017_rmvnMicroBeh}
\title{Fogel2017_rmvnMicroBeh: Generates a random sample selected from the multivariate normal distribution based on Fogel et al., 2017}
\usage{
Fogel2017_rmvnMicroBeh(n)
}
\arguments{
\item{n}{number random draws from the random multivariate normal distribution}
}
\value{
A simulated data set of microstructure behaviors
}
\description{
This function selects a random sample from the multivariate normal distribution that includes: model parameters, total intake, and number of bites.
}
\examples{

\dontrun{
}


}
