% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPM_Fit_r_reparam.R
\name{FPM_Fit_r_reparam}
\alias{FPM_Fit_r_reparam}
\title{FPM_Fit_r_reparam: Fits parameters for the First Principles Model with r re-parameterized to be e^r = ln(r)}
\usage{
FPM_Fit_r_reparam(data, parameters, timeVar, intakeVar, Emax)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric parameters for the First Principles Model in the format: c(theta, r)}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
A list with of values from optim fit
\itemize{
  \item{par}{fitted parameters theta (par[1]) and r (par[2])}
  \item{ value}{-2 log-likelihood corresponding to the fitted parameters}
  \item{counts}{the number of itterations to achieve fit}
  \item{convergence}{code for convergence status; see optim}
  \item{message}{a string with additional information, generally NULL}
}
}
\description{
This function fits theta and r for the First Principles Model of the
the cumulative intake curves from Thomas et al., (2017) using
optim {stats}. r re-parameterized to be e^r = ln(r)
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Fit}}.
}
