% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPM_n2ll.R
\name{FPM_n2ll}
\alias{FPM_n2ll}
\title{FPM_n2ll: Calculates -2 loglikelihood for the First Principles Model}
\usage{
FPM_n2ll(data, par, timeVar, intakeVar, Emax)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{par}{A set of numeric parameters for the First Principles Model in the format: c(theta, r)}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
NEED TO EDIT
}
\description{
This function calculates the -2 loglikelihood for specific function for
the First Principles Model of the the cumulative intake curve (Thomas et al., 2017)
}
\examples{

\dontrun{
}

}
\seealso{
To get best fit parameters for the First Principles Model use \code{\link{FPM_Fit}}.
To get fit your intake data using the  Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Fit}}
and \code{\link{Kissileff_n2ll}}.
}
