% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPM_Time.R
\name{FPM_Time}
\alias{FPM_Time}
\title{FPM_Time: Estimates the elapse time using the First Principles Model}
\usage{
FPM_Time(intake, parameters, Emax)
}
\arguments{
\item{intake}{The cumulative intake since meal start.}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
Numeric value indicating the time since start of meal for given cumulative.
}
\description{
This function estimates the elapsed time since start of a meal for a given cumulative intake
using the First Principles Model for the cumulative intake curves from Thomas et al., (2017).
}
\examples{
#Get the time when 15 grams have been consumed:
FPM_Time(15, c(30, .25), 300)

#save \theta and r as an object first:
beta_coefs = c(30, .25)
FPM_Time(15, beta_coefs, 300)

\dontrun{
#be careful of how you format the list of \beta coefficients. These are incorrect:
FPM_Time(15, (30, .25), 300)
FPM_Time(15, 30, .25, 300)
}

}
\seealso{
For the reverse calculation, see \code{\link{FPM_Intake}} to get meal cumulative intake
at a given time. To get cumulative intake and meal time using Kisslieff's
quadratic model (Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Intake}}
and \code{\link{Kissileff_Time}}.
}
