% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kissileff_Fit.R
\name{Kissileff_Fit}
\alias{Kissileff_Fit}
\title{Kissileff_Fit: Fits \beta coeffiecents for Kissileff's quadratic model}
\usage{
Kissileff_Fit(data, parameters, timeVar, intakeVar)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}
}
\value{
NEED TO EDIT
}
\description{
This function fist the \beta coefficients for Kissileff's quadratic model
for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001) using
optim {stats}.
}
\section{}{

}

\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the First Principles Model (Thomas et al., 2017),
see \code{\link{FPM_FIT}}.

Other aggregate functions: 
\code{\link{FPM_Fit}()},
\code{\link{FPM_Intake}()},
\code{\link{FPM_Time}()},
\code{\link{Kissileff_Intake}()},
\code{\link{Kissileff_Time}()},
\code{\link{simBites}()}
}
\concept{aggregate functions}
