% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kissileff_n2ll.R
\name{Kissileff_n2ll}
\alias{Kissileff_n2ll}
\title{Kissileff_n2ll: Calculates -2 loglikelihood for Kissileff's quadratic model}
\usage{
Kissileff_n2ll(data, par, timeVar, intakeVar)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{par}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic)}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}
}
\value{
The -2 log-likelihood for the model given the specified parameters.
}
\description{
This function calculates the -2 loglikelihood for specific function for
Kissileff's quadratic model for the cumulative intake curves
(Kissileff, 1982; Kissileff & Guss, 2001)
}
\examples{

\dontrun{
}

}
\seealso{
To get best fit parameters for the First Principles Model use \code{\link{Kissileff_Fit}}.
To get fit your intake data using the  Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{FPM_Fit}}
and \code{\link{FPM_n2ll}}.
}
