% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kissileff_Time.R
\name{Kissileff_Time}
\alias{Kissileff_Time}
\title{Kissileff_Time: Estimates elapsed time using Kisslieff's quadratic model}
\usage{
Kissileff_Time(intake, parameters, message = TRUE)
}
\arguments{
\item{intake}{The cumulative intake since meal start.}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{message}{(optional) Return warning message about not being able to fit a time point. Default is TRUE}
}
\value{
Numeric value indicating the time since start of meal for given cumulative intake and parameters.
}
\description{
This function estimates the elapsed time since start of a meal for a given cumulative intake using
the quadratic model for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001).
}
\examples{
#Get the time when 15 grams have been consumed:
Kissileff_Intake(15, c(10, 7, -0.3))

#save \beta coefficents as an object first:
beta_coefs = c(10, 7, -0.3)
Kissileff_Intake(15, beta_coefs)

\dontrun{
#be careful of how you format the list of \beta coefficients. These are incorrect:
Kissileff_Time(15, (10, 7, -.3))
Kissileff_Time(15, 10, 7, -.3)
}

}
\seealso{
For the reverse calculation, see \code{\link{Kissileff_Intake}} to get cumulative intake
from entered time since start of meal. To get cumulative intake and meal time using the First
Principles Model (Thomas et al., 2017), see \code{\link{FPM_Intake}} and \code{\link{FPM_Time}}.
}
