% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearBites_Fogel2017.R
\docType{data}
\name{linearBites_Fogel2017}
\alias{linearBites_Fogel2017}
\title{linearBites_Fogel2017 simulated bite data from an exponential distribution}
\format{
A data frame with 500 rows and 14 variables:
\describe{
    \item{id}{Simulation number}
    \item{Bite}{Bite number}
    \item{SampledTime}{Time sampled from a truncated exponential distribution}
    \item{EstimatedCumulativeIntake}{Cumulative intake based on average bite size}
}
}
\usage{
data(linearBites_Fogel2017)
}
\description{
Simulated data based on Fogel et al., 2017 child microstructure behavior means and variance structure. 500 sets of mictrostructure behaviors were randomly selected from a multivariate normal distribution. Bites were generated using randomly sampling bite timings using a linear distribution. See generation script DataGen_SimDat_Fogel2017.R in raw-data.
}
\examples{
data(linearBites_Fogel2017)
nBites <- attr(linearBites_Fogel2017, "Bites")

}
\references{
Fogel A, Goh AT, Fries LR, et al. A description of an “obesogenic” eating style that promotes higher energy intake and is associated with greater adiposity in 4.5year-old children: Results from the GUSTO cohort. Physiology & Behavior. 2017;176:107-116. doi:10.1016/j.physbeh.2017.02.013

Fogel A, Goh AT, Fries LR, et al. Physiology & Behavior. 2017;176:107-116
(\href{https://pubmed.ncbi.nlm.nih.gov/28213204/}{PubMed})
}
\keyword{datasets}
