% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LODE_Fit.R
\name{LODE_Fit}
\alias{LODE_Fit}
\title{LODE_Fit: Fits parameters for the Logistic Ordinary Differential Equation (LODE) Model}
\usage{
LODE_Fit(data, parameters, timeVar, intakeVar, Emax)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric parameters for the Logistic Ordinary Differential Equation (LODE) in the format: c(theta, r)}

\item{timeVar}{Name of the variable in data that contains timing data}

\item{intakeVar}{AName of the variable in data that contains cumulative intake data}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
A list with of values from optim fit
\itemize{
  \item{par}{fitted parameters theta (par[1]) and r (par[2])}
  \item{ value}{-2 log-likelihood corresponding to the fitted parameters}
  \item{counts}{the number of itterations to achieve fit}
  \item{convergence}{code for convergence status; see optim}
  \item{message}{a string with additional information, generally NULL}
}
}
\description{
This function fits theta and r for the Logistic Ordinary Differential Equation (LODE) Model of the
the cumulative intake curves from Thomas et al., (2017) using
optim {stats}.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Quad_Fit}}.
}
