% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LODEincorrect_Intake.R
\name{LODEincorrect_Intake}
\alias{LODEincorrect_Intake}
\title{LODEincorrect_Intake: Estimates cumulative intake using the Logistic Ordinary Differential Equation (LODE) Model as written in the uncorrected Thomas et al., (2017) paper.}
\usage{
LODEincorrect_Intake(time, parameters, Emax)
}
\arguments{
\item{time}{The time in minutes since meal start.}

\item{parameters}{A set of numeric parameters for the LODE model in the format: c(theta, r)}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
Numeric value indicating the cumulative intake at specified time given the specified parameters.
}
\description{
This function estimates cumulative intake at a given time since start of meal using
the Logistic Ordinary Differential Equation (LODE) Model for the cumulative intake curves from Thomas et al., (2017), with the incorrect formula from the uncorrected paper.
}
\examples{
#Get cumulative intake at minute three:
LODE_Intake(3, c(30, .25), 300)

#save \theta and r as an object first
params = c(30, .25)
LODE_Intake(3, params, 300)

\dontrun{
#be careful of how you format the parameter list. These are incorrect:
LODE_Intake(3, (30, .25), 300)
LODE_Intake(3, 30, .25, 300)
}

}
\seealso{
For the reverse calculation, see \code{\link{LODE_Time}} to get meal time
from entered cumulative intake. To get cumulative intake and meal time using the
Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Quad_Intake}}
and \code{\link{Quad_Time}}.
}
