% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LODEincorrect_Time.R
\name{LODEincorrect_Time}
\alias{LODEincorrect_Time}
\title{LODEincorrect_Time: Estimates the elapse time using the Logistic Ordinary Differential Equation (LODE) Model as written in the uncorrected Thomas et al., (2017) paper.}
\usage{
LODEincorrect_Time(intake, parameters, Emax, message = TRUE)
}
\arguments{
\item{intake}{The cumulative intake since meal start.}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{Emax}{The total cumulative intake at the end of the meal.}

\item{message}{(optional) Return warning message about not being able to fit a time point. Default is TRUE}
}
\value{
Numeric value indicating the time since start of meal for given cumulative intake and parameters.
}
\description{
This function estimates the elapsed time since start of a meal for a given cumulative intake
using the Logistic Ordinary Differential Equation (LODE) Model for the cumulative intake curves from Thomas et al., (2017), with the incorrect formula from the uncorrected paper.
}
\examples{
#Get the time when 15 grams have been consumed:
LODE_Time(15, c(30, .25), 300)

#save \theta and r as an object first:
params = c(30, .25)
LODE_Time(15, params, 300)

\dontrun{
#be careful of how you format the vector of parameters. These are incorrect:
LODE_Time(15, (30, .25), 300)
LODE_Time(15, 30, .25, 300)
}

}
\seealso{
For the reverse calculation, see \code{\link{LODE_Intake}} to get meal cumulative intake
at a given time. To get cumulative intake and meal time using the
Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Quad_Intake}}
and \code{\link{Quad_Time}}.
}
