% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDistinctness.R
\name{modelDistinctness}
\alias{modelDistinctness}
\title{modelDistinctness: Estimates the distinctness of an estimated parameter relative to all other estimate confidence bounds in dataset}
\usage{
modelDistinctness(data, paramVar, upperVar, lowerVar, sep, cutoff = 0.05, id)
}
\arguments{
\item{data}{A dataset containing all model parameter estimates and their confidence bounds}

\item{paramVar}{Variable name  for recovered parameter estimate}

\item{upperVar}{Variable name OR prefix for upper confidence bounds associated with estimate.}

\item{lowerVar}{Variable name OR prefix for lower confidence bounds associated with estimate.}

\item{sep}{(optional) If upperVar and lowerVar are prefixes, then need to enter a separator character to use. Ex., if paramVar = 'theta' and upperVar = 'u95CI', a sep = '_' would result in upperVar being 'u95CI_theta'.}

\item{cutoff}{The proportion of overlap that serves as a cutoff for distinctness. Default is 0.05, which would consider 5 percent or less overlap distinct}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A list with two values
    \item{nOverlap}{number of other observations that the estimate is distinct from}
    \item{Distinct}{whether estimate meets cutoff for dinstinctness}
}
\description{
This function the distinctness of each estimate by comparing it to the confidence bounds of all other estimates. An estimate is more pdistinct if it does not fall into other estimates confidences bounds.
}
\examples{

\dontrun{

}

}
