% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelError.R
\name{modelError}
\alias{modelError}
\title{modelError: Returns error metrics for predicted cumulative intake or timing}
\usage{
modelError(
  data,
  parameters,
  timeVar,
  intakeVar,
  model_str = "LODE",
  method,
  error_measure = "intake",
  adjustNA = "interpolate"
)
}
\arguments{
\item{data}{The dataset with 'True' or observed bite timing and cumulative intake.}

\item{parameters}{Fitted parameters - The Quadratic model needs an intercept, linear slope, and quadratic slope entered in that order and Logistic Ordinary Differential Equation (LODE) Model needs theta and r entered in that order}

\item{timeVar}{The variable name for the 'True' or observed bite timing in dataset}

\item{intakeVar}{The variable name for the 'True' or observed cumulative intake in dataset}

\item{model_str}{(optional) Only needed if simulating data. 'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'}

\item{method}{Specify which error metric to compute - 'rmse' will calculated root mean squared error, 'R2' will calculate a pseudo-R^2, 'both' with return both measures}

\item{error_measure}{Which variable to use to calculate error - 'timing' will use bite timing, 'intake' will use cumulative intake, 'both' will return both. Default is 'intake'.}

\item{adjustNA}{The method used to adjust predicted values that are not real integers, resulting in NA values. 'interpolate' will average the true data values from adjacent bites while 'minmax' will use the minimum or maximum true data values for bites at the start and end of the meal, respectively. Default is 'interpolate'.}
}
\value{
A list with 3-4 values depending on input arguments
    \item{rmse}{the root mean squared error (if requested)}
    \item{R2}{pseudo-R^2 (if requested)}
    \item{nNAs}{number of timepoints where the value (time or intake) could not be estimated}
    \item{nNeg}{number of timepoints where the value (time or intake) was predicted to be negative}
}
\description{
This function calculates the root mean squared error for either bites or timing based on recovered data.
}
\examples{

\dontrun{

}

}
