% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2LL_gridsearch.R
\name{n2LL_grid}
\alias{n2LL_grid}
\title{n2LL_grid: Completes a grid search of the -2 log-likelihood space by varying parameter values}
\usage{
n2LL_grid(
  data,
  parameters = c(10, 0.1),
  model_str = "FPM",
  paramStep,
  bounds,
  nSteps = 10,
  type = "linear",
  timeVar,
  intakeVar,
  graph = TRUE
)
}
\arguments{
\item{data}{A bite dataset containing cumulative intake and timing for each bite}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c(10, 1, -1)) and FPM_Time needs 2 starting parameters (default is c(10, .10)).}

\item{model_str}{The base model to use--'FPM' for the first principles model and 'Kissileff' for the quadratic model.
Default is 'FPM'.}

\item{paramStep}{Names of parameters that will be varied during calculation of -2 log-likelihoods: c('theta', 'r)}

\item{bounds}{Upper and lower bounds for parameter values. If varying the value of multiple parameters (i.e., more than one listed in paramStep), this will take a list of boundary vectors: boundsList = list(c(10, 15), c(0, 1)). THe order of this must match the order of the parameters specified in paramStep}

\item{nSteps}{Number of steps to vary each parameter. Note: this is not the number of -2 log-likihood estimates}

\item{type}{The type of search to complete. 'grid' will calculate the -2 log-likihood for every combination of parameter step value combinations (default). 'linear' will calculate -2 log-likehood for every step of a single parameter while holding other parameters constant at the value entered in parameters.}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{graph}{(optional) A logical indicator indicate whether a graph should be returned. Default is TRUE}
}
\value{
A dataset with the -2 log-likelihood and parameter values for each step of the search. Optionally, will also return a graph of the likelihood space.
}
\description{
This function computes -2 log-likelihood values while varying parameter values.
}
\examples{

\dontrun{
}

}
\seealso{
This function relies on \code{\link{simBites}} and \code{\link{IntakeModelParams}}.
}
