% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramCheck.R
\name{paramCheck}
\alias{paramCheck}
\title{paramCheck: checks the feasibility of parameters recovered from bite data}
\usage{
paramCheck(Emax, parameters, model_str = "LODE", id)
}
\arguments{
\item{Emax}{A numeric value for total cumulative intake.}

\item{parameters}{A set of numeric parameters: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order (default is c(10, 1, -1)) and Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order (default is c(10, .10)).}

\item{model_str}{The base model to use--'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'.}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A boolean operator indicating if the provided parameters reflect a feasible cumulative intake curve (TRUE) or not (FALSE)
}
\description{
This function checks the feasibility parameters recovered from bite data for the Quadratic and Logistic Ordinary Differential Equation (LODE) model. For the Quadratic model, it checks the vertices relative to the sign of the quadratic coefficient and total intake. For the LODE model, it checks the r limit.
}
\examples{

\dontrun{
}

}
