% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRecovery.R
\name{ParamRecovery}
\alias{ParamRecovery}
\title{ParamRecovery: This function recovers parameters for cumulative intake model and the bite data provided}
\usage{
ParamRecovery(
  nBites,
  Emax,
  parameters,
  model_str = "LODE",
  procNoise = TRUE,
  measureNoise = FALSE,
  pNoise_biteSizeSD = NA,
  mNoise_biteTimeSD = NA,
  mNoise_biteSizeCat = "mean",
  keepBites = FALSE,
  conf = 95,
  rmse = NA
)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{parameters}{A set of numeric parameters: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order (default is c(10, 1, -1)) and Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order (default is c(10, .10)).}

\item{model_str}{The base model to use--'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'.}

\item{procNoise}{(optional) A logical indicator for adding random process noise to the bite data by jittering bite size with bite timing estimated from jittered bite sizes. This uses the default jitter amount (smallest distance/5). Default value is TRUE.}

\item{measureNoise}{(optional) A logical indicator for measurement noise after the estimation of bite timings. Default value is FALSE.}

\item{pNoise_biteSizeSD}{(optional) This allows you to enter the standard deviation of individuals bites sizes and will replace the default procNoise routine (jittered bite sizes). Bite sizes will be randomly chosen from a normal distribution truncated at min = 0 with mean = Emax/nBites and standard deviation equal to the entered value. procNoise must be set to TRUE, otherwise this argument will be ignored.}

\item{mNoise_biteTimeSD}{(optional) This allows you to enter the standard deviation for adjusting bite timing and will replace the default (jittered bite timing). The noise add to each timepoint will be chosen from a normal distribution  with mean = 0 and standard deviation entered. measureNoise must be set to to 'BiteTiming' or 'Both' otherwise this argument will be ignored. Note: the normal distribution will be truncated at at each timepoint so that the time for timepoint t is not less than timepoint t-1.}

\item{mNoise_biteSizeCat}{(option) This allows you to alter the default for bite size error (average bite size) by
entering category cut points or NA to skip this measurement error. Cut points must equal n - 1 categories (e.g., if want three categories you would enter the small-medium and medium-large large cut/boundry points). Cut points will be left/lower inclusive but exclude upper boundary. Bite sizes within each category will be set to the average bite size for that category. This will replace the default measureNoise routine (all bites = average bite size). measureNoise must be set to to 'BiteSize' or 'Both' otherwise this argument will be ignored.}

\item{keepBites}{(optional) This is a logical indicator for whether to return the simulated bite dataset with
elapsed time and cumulative intake for each bite across all simulated intake curves. The returned cumulative
intake data will use the same bite timing as in the initial data but will estimate intake for each bite based on
the recovered model parameters. Default is FALSE.}

\item{conf}{Numeric value for the percent confidence desired. Default is 95, which provides 95\% confident
intervals. If no confidence intervals are desired, set conf to FALSE.}

\item{rmse}{(optional) A string indicating which measure to compute root mean square error for. Options include:
'timing' for bite timing, 'intake' for cumulative intake', 'both' to compute for both timing and intake.
If not specified, rmse is not computed. Default is to not compute rmse.}
}
\value{
Either 1 or 2 datasets. It will always return a dataset with recovered parameters but will only return a list with bite data sets for each simulation if keepBites = TRUE
}
\description{
This function simulates the cumulative intake curve using average bites size and then fitting
the model parameters for each curve. Process noise can be used rather than average bite size, if wanted.
Additionally, measurement error can be added after the estimation of bite timing (from bite size) by reverting
to average bite size or categorizing bite sizes and jittering the bite timing. The distinction between processes
and measurement noise is that process noise is added before the calculation of bite timing while measurement noise
is added after and there is no adjustment to fit the model. The parameters will be fit using either the
Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001) or the Logistic Ordinary Differential
Equation (LODE) Model (Thomas et al., 2017), total intake (Emax), and number of bites.
}
\examples{

\dontrun{
}

}
\seealso{
This function relies on \code{\link{n2LL_LODE}} and \code{\link{n2LL_Quad}}.
}
