% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quad_Intake.R
\name{Quad_Intake}
\alias{Quad_Intake}
\title{Quad_Intake: Estimates cumulative intake using the Quadratic model}
\usage{
Quad_Intake(time, parameters)
}
\arguments{
\item{time}{The time in minutes since meal start.}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}
}
\value{
Numeric value indicating the cumulative intake at specified time for the given set of parameters.
}
\description{
This function estimates cumulative intake at a given time since start of meal using the quadratic model for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001).
}
\examples{
#Get cumulative intake at minute three:
Quad_Intake(3, c(10, 7, -0.3))

#save \beta coefficents as an object first:
beta_coefs = c(10, 7, -0.3)
Quad_Intake(3, beta_coefs)

\dontrun{
#be careful of how you format the list of \beta coefficients. These are incorrect:
Quad_Intake(3, (10, 7, -.3))
Quad_Intake(15, 10, 7, -.3)
}

}
\seealso{
For the reverse calculation, see \code{\link{Quad_Time}} to get meal time from entered cumulative intake. To get cumulative intake and meal time using the Logistic Ordinary Differential Equation (LODE) Model(Thomas et al., 2017), see \code{\link{LODE_Intake}} and \code{\link{LODE_Time}}.
}
