% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quad_Time.R
\name{Quad_Time}
\alias{Quad_Time}
\title{Quad_Time: Estimates elapsed time using the Quadratic model}
\usage{
Quad_Time(intake, parameters, message = TRUE)
}
\arguments{
\item{intake}{The cumulative intake since meal start.}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{message}{(optional) Return warning message about not being able to fit a time point. Default is TRUE}
}
\value{
Numeric value indicating the time since start of meal for given cumulative intake and parameters.
}
\description{
This function estimates the elapsed time since start of a meal for a given cumulative intake using
the quadratic model for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001).
}
\examples{
#Get the time when 15 grams have been consumed:
Quad_Intake(15, c(10, 7, -0.3))

#save \beta coefficents as an object first:
beta_coefs = c(10, 7, -0.3)
Quad_Intake(15, beta_coefs)

\dontrun{
#be careful of how you format the list of \beta coefficients. These are incorrect:
Quad_Time(15, (10, 7, -.3))
Quad_Time(15, 10, 7, -.3)
}

}
\seealso{
For the reverse calculation, see \code{\link{Quad_Intake}} to get cumulative intake
from entered time since start of meal. To get cumulative intake and meal time using the
Logistic Ordinary Differential Equation (LODE) Model (Thomas et al., 2017),
see \code{\link{LODE_Intake}} and \code{\link{LODE_Time}}.
}
