% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{RMSE: Calculates the root mean squared error}
\usage{
RMSE(trueVal, predVal)
}
\arguments{
\item{trueVal}{Vector of 'True' values}

\item{predVal}{Vector of predicted values}
}
\value{
NEED TO EDIT
}
\description{
This function calculates the root mean squared error between true and predicted values
}
\examples{

\dontrun{
}


}
