% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSEcalc.R
\name{RMSEcalc}
\alias{RMSEcalc}
\title{RMSEcalc: Wrapper of model fit functions that calculates the root mean squared error}
\usage{
RMSEcalc(
  data,
  parameters,
  timeVar,
  intakeVar,
  model_str = "LODE",
  error_outcome = "timing",
  Emax
)
}
\arguments{
\item{data}{The dataset with 'True' bite timing and cumulative intake.}

\item{parameters}{Fitted parameters for which error is being tested. The Quadratic model needs an intercept, linear slope, and quadratic slope entered in that order and Logistic Ordinary Differential Equation (LODE) Model needs theta and r entered in that order}

\item{timeVar}{The variable name for the 'True' bite timing in dataset}

\item{intakeVar}{The variable name for the 'True' cumulative intake in dataset}

\item{model_str}{The base model to use--'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'.}

\item{error_outcome}{Which variable to use to calculate error - 'timing' will use bite timing and 'intake' will use cumulative intake. Default is 'timing'.}

\item{Emax}{A numeric value that represents total cumulative intake.}
}
\value{
A list with two values
    \item{rmse}{the root mean squared error}
    \item{nNAs}{number of timepoints where the value (time or intake) could not be estimated}
}
\description{
This function wraps the RMSE functions to calculate the root mean squared error for either bites or timing.
}
\examples{

\dontrun{
}

}
