% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBitesExp.R
\name{simBitesExp}
\alias{simBitesExp}
\title{simBitesExp: Generates a bite dataset by randomly sampling time from a exponential decay distribution}
\usage{
simBitesExp(mealdur, nBites, Emax, id)
}
\arguments{
\item{mealdur}{Meal duration in minutes}

\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A bite dataset generated by randomly sampling time
}
\description{
This function generates a bite dataset from meal duration, number of bites, and total intake (Emax) by randomly sampling a exponential decay distribution
}
\examples{

\dontrun{
}


}
