% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBitesLogit.R
\name{simBitesLogit}
\alias{simBitesLogit}
\title{simBitesLogit: Generates a bite dataset by randomly sampling time}
\usage{
simBitesLogit(mealdur, nBites, Emax, id)
}
\arguments{
\item{mealdur}{Meal duration in minutes}

\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{id}{(optional) ID number, if exists will be added to output_dat. Default is no id.}
}
\value{
A bite dataset generated by randomly sampling time
}
\description{
This function generates a bite dataset from meal duration, number of bites, and total intake (Emax) by randomly sampling a logistic distrubtion truncated to only include positive values
}
\examples{

\dontrun{
}


}
