% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBitesQuad.R, R/simBitesUQuad.R
\name{simBitesQuad}
\alias{simBitesQuad}
\title{simBitesQuad: Generates a bite dataset by randomly sampling time from a truncated quadratic distribution}
\usage{
simBitesQuad(mealdur, nBites, Emax, id)

simBitesQuad(mealdur, nBites, Emax, id)
}
\arguments{
\item{mealdur}{Meal duration in minutes}

\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A bite dataset generated by randomly sampling time

A bite dataset generated by randomly sampling time
}
\description{
This function generates a bite dataset from meal duration, number of bites, and total intake (Emax) by randomly sampling a quadratic distribution truncated to only include positive values

This function generates a bite dataset from meal duration, number of bites, and total intake (Emax) by randomly sampling a u-quadratic distribution truncated to only include positive values
}
\examples{

\dontrun{
}



\dontrun{
}


}
