% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biteMeasureNoise.R
\name{biteMeasureNoise}
\alias{biteMeasureNoise}
\title{ParamRecovery: This function recovers parameters for cumulative intake model and the bite data provided}
\usage{
biteMeasureNoise(
  BiteDat,
  nBites,
  Emax,
  TimeVar = "EstimatedTime",
  BiteVar = "BiteGrams",
  measureNoise = FALSE,
  mNoise_biteTimeSD = NA,
  mNoise_biteSizeCat = "mean"
)
}
\arguments{
\item{BiteDat}{A dataset with Bites, bite sizes, cumulative intake, and bite timing.}

\item{nBites}{A numeric value for total number of bites in a meal.}

\item{Emax}{A numeric value for total cumulative intake.}

\item{TimeVar}{String reflecting name for bite timing variable BiteDat}

\item{BiteVar}{String reflecting name for bite size variable in BiteDat dataset}

\item{measureNoise}{(optional) A string indicating they type of measurement noise to add. The options include:
'BiteSize' - will use average bite size for parameter recovery; 'BiteTiming' - add noise to bite timing (jittered); or 'Both' - will apply both types of measurement noise. This noise is applied to bite data after initial parameterization and before parameter recovery. Default is no measurement error.}

\item{mNoise_biteTimeSD}{(optional) This allows you to enter the standard deviation for adjusting bite timing and will replace the default (jittered bite timing). The noise add to each timepoint will be chosen from a normal distribution  with mean = 0 and standard deviation entered. measureNoise must be set to to 'BiteTiming' or 'Both' otherwise this argument will be ignored. Note: the normal distribution will be truncated at at each timepoint so that the time for timepoint t is not less than timepoint t-1.}

\item{mNoise_biteSizeCat}{(option) This allows you to alter the default for bite size error (average bite size) by
entering category cut points or NA to skip this measurement error. Cut points must equal n - 1 categories (e.g., if want three categories you would enter the small-medium and medium-large large cut/boundry points). Cut points will be left/lower inclusive but exclude upper boundary. Bite sizes within each category will be set to the average bite size for that category. This will replace the default measureNoise routine (all bites = average bite size). measureNoise must be set to to 'BiteSize' or 'Both' otherwise this argument will be ignored.}
}
\value{
It will always return a dataset with adjusted bite timing reflecting measurement error
}
\description{
This function simulates the cumulative intake curve using average bites size and then fitting
the model parameters for each curve. Process noise can be used rather than average bite size, if wanted.
Additionally, measurement error can be added after the estimation of bite timing (from bite size) by reverting
to average bite size or categorizing bite sizes and jittering the bite timing. The distinction between processes
and measurement noise is that process noise is added before the calculation of bite timing while measurement noise
is added after and there is no adjustment to fit the model. The parameters will be fit using either the
Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001) or the Logistic Ordinary Differential Equation
(LODE) Model (Thomas et al., 2017), total intake (Emax), and number of bites.
}
\examples{

\dontrun{
}

}
\seealso{
To add process noise to bite data *before* bite timing calculation see \code{\link{biteProcNoise}}
}
