% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biteProcNoise.R
\name{biteProcNoise}
\alias{biteProcNoise}
\title{biteProcNoise: Adds process noise to cumulative intake by bites}
\usage{
biteProcNoise(cumulativeBites, nBites, Emax, Et0 = 0, pNoiseSD = NA)
}
\arguments{
\item{nBites}{A numeric value for total number of bites in a meal.}

\item{Emax}{A numeric value for total cumulative intake.}

\item{Et0}{Only needed if data entered is cumulative intake. Minimum intake in grams that maintains a time >= 0 (note, for LODE this is always 0, for the Quadratic models, need to calculate relative to the sign of the quadratic parameter and the vertex of the quadratic equation. See @biteIntake for full explanation).}

\item{pNoiseSD}{(optional) This allows you to enter the standard deviation of individuals bites sizes and will replace the default procNoise routine (jittered bite sizes). Bite sizes will be randomly chosen from a normal distribution truncated at min = 0 with mean = Emax/nBites and standard deviation equal to the entered value. procNoise must be set to TRUE, otherwise this argument will be ignored.}

\item{data}{A numeric vector representing the cumulative intake at each bite or bite timing}

\item{mealDur}{Only needed if data entered is bite timing. Max duration of the meal.}
}
\value{
A dataset of bites and cumulative intake with process noise added
}
\description{
This function generates a bite data set that includes cumulative intake for *n* bites
after adding process noise.
}
\details{
If sd_bitesize is specified, the bites size will randomly vary across the meal using a Gussian
ditribution with mean = average bite size and SD = sd_bitesize.
}
\examples{

\dontrun{
}

}
\seealso{
To add measurement noise to bite data *after* bite timing calculation see \code{\link{biteMeasureNoise}}
}
