% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biteProcNoise.R
\name{biteProcNoise}
\alias{biteProcNoise}
\title{biteProcNoise: Adds process noise to cumulative intake by bites}
\usage{
biteProcNoise(cumulativeBites, nBites, Emax, minE = 0, pNoise_biteSizeSD = NA)
}
\arguments{
\item{cumulativeBites}{A numeric vector representing the cumulative intake at each bite}

\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{minE}{A numeric value that represents the minimum intake in grams that maintains a time >= 0 (note, for FPM this is always 0, for Kissileff quadratic models, need to calculate relative to the sign of the quadratic parameter and the vertex of the quadratic equation. See @simBites for full explanation)}

\item{pNoise_biteSizeSD}{(optional) This allows you to enter the standard deviation of individuals bites sizes and will replace the default procNoise routine (jittered bite sizes). Bite sizes will be randomly chosen from a normal distribution truncated at min = 0 with mean = Emax/nBites and standard deviation equal to the entered value. procNoise must be set to TRUE, otherwise this argument will be ignored.}
}
\value{
A dataset of bites and cumulative intake with process noise added
}
\description{
This function generates a bite data set that includes cumulative intake for *n* bites
after adding process noise.
}
\details{
If sd_bitesize is specified, the bites size will randomly vary across the meal using a Gussian
ditribution with mean = average bite size and SD = sd_bitesize.
}
\examples{

\dontrun{
}

}
\seealso{
To add measurement noise to bite data *after* bite timing calculation see \code{\link{biteMeasureNoise}}
}
