% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biteProcNoise.R
\name{biteProcNoise}
\alias{biteProcNoise}
\title{biteProcNoise: Adds process noise to either bite intake or bite timing}
\usage{
biteProcNoise(
  data,
  type,
  Emax,
  mealDur,
  model_str = "LODE",
  parameters,
  pNoiseSD = NA
)
}
\arguments{
\item{data}{Either a vector of cumulative intake at each bite or bite timings}

\item{type}{String indicating type of data: 'intake' or 'time'}

\item{Emax}{(optional) Total cumulative intake. If not entered, will use max of data for type = 'intake'}

\item{mealDur}{(optional) Meal duration in minutes. If not entered will use max of data entered for type = 'time'}

\item{model_str}{(optional) Only needed if simulating data. 'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'}

\item{parameters}{(optional) Only needed if model_str = 'Quad'. A set of numeric parameters; the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order.}

\item{pNoiseSD}{(optional) Only use if want to replace default approach (jitter). Standard deviation of individuals to use to generate a Gaussian distribution. Mean will be set depending on entered type: intake - mean bite size; time - mean inter-bite-interval. Distribution will be truncated at 0 if needed.}
}
\value{
If type = 'timing', a vector of bite timings with process noise added. If type = 'intake', a dataset of bite sizes and cumulative intake with process noise added.
}
\description{
Simulates process noise to better approximate human meal microstructure data. Process error is simulated by adding noise to bite intake and timing so each bite size differs slightly and the inter-bite-interval varies across meal time.
}
\details{
If pNoiseSD is specified, the bite sizes will randomly vary across the meal using a Gaussian distribution. When adding process noise to intake data, the distribution will have a mean = average bite size and SD = pNoiseSD. For the Quadratic model, the distribution will be truncated at the minimum bite size for time = 0 give the entered parameters. For the LODE model, it will be truncated at 0. When adding process noise to time data, the distribution will have a mean = average time between bites and SD = pNoiseSD.
}
\examples{
#simulate bite dataset
bite_data <- biteIntake(nBites = 15, Emax = 300, parameters = c(10, .10))

#add noise to data
intake_noise <- biteProcNoise(data = bite_data$CumulativeGrams_procNoise, type = 'intake')

timing_noise <- biteProcNoise(data = bite_data$EstimatedTime_procNoise, type = 'time')

\dontrun{
}

}
\seealso{
To add measurement noise to bite data *after* initial parametarization see \code{\link{biteMeasureNoise}}
}
