% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_LRTest.R
\name{CI_LRTest}
\alias{CI_LRTest}
\title{CI_LRTest: Wrapper of model fit functions that calculates the likelihood ratio test}
\usage{
CI_LRTest(
  data,
  par,
  model_str = "LODE",
  timeVar,
  intakeVar,
  min_n2ll,
  paramIndex,
  conf = 95,
  bound
)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{par}{A set of numeric parameters: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order; the Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order}

\item{model_str}{The base model to use--'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'.}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{min_n2ll}{The minimum -2 log-likelihood value obtained when fitting the parameters.}

\item{paramIndex}{The index number for par that corresponds to the parameter the CI is being fit for. E.g., if LODE Model, par[1] would be theta and par[2] would be r.}

\item{conf}{Numeric value for the percent confidence desired. Default is 95 for the 95th percent CI.}

\item{bound}{A string with the boundary value desired: 'upper' or 'lower'}
}
\value{
The likelihood ratio test for the CI bound and value (upper v lower) requested
}
\description{
This function warps model fit functions called through optim {stats} to calcualte the likelihood ratio test. This is a function that, when called by optim {stats} will identify a confidence bound of a parameter by optim's minimization of the calculated LRT value.
}
\examples{

\dontrun{
}

}
