% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPM_Fit.R
\name{FPM_Fit}
\alias{FPM_Fit}
\title{FPM_Fit: Fits parameters for the First Principles Model}
\usage{
FPM_Fit(data, parameters, timeVar, intakeVar, Emax)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric parameters for the First Principles Model in the format: c(theta, r)}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{Emax}{The total cumulative intake at the end of the meal.}
}
\value{
NEED TO EDIT
}
\description{
This function fits theta and r for the First Principles Model of the
the cumulative intake curves from Thomas et al., (2017) using
optim {stats}.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Fit}}.
}
