% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBiteDat.R
\name{genBiteDat}
\alias{genBiteDat}
\title{genBiteDat: Generates a cumulative intake bite dataset from a specific distribution}
\usage{
genBiteDat(
  nBites,
  Emax,
  mealDur,
  genDist = "logis",
  return_params = FALSE,
  id = NA
)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{mealDur}{Meal duration in minutes}

\item{genDist}{A string or vector of strings for the generating distribution. Options include: 'logis', 'quad', 'u-quad', 'exp', or 'linear'. Default is 'logis'. Can enter more than one to return different Time variables.}

\item{return_params}{(optional) A boolean indicating if model parameters should be returned. Only relevant if a model_str is entered. Default is FALSE.}

\item{id}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A bite dataset with bite timing, bite size, and cumulative intake for each bite and a set of fitted parameters
}
\description{
This function generates a bite dataset based on a specified distribution and model.
Using total meal duration and nBites, bite timing is randomly sampled from the specified distribution.
Cumulative intake is generated using average bite size. The returned cumulative intake dataset will be
checked to ensure it is a feasible intake pattern according to the LODE and quadratic models.
}
\examples{

\dontrun{
}


}
