% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntakeModelParams.R
\name{IntakeModelParams}
\alias{IntakeModelParams}
\title{IntakeModelParams: Fits model parameters for cumulative intake curves}
\usage{
IntakeModelParams(
  data,
  parameters,
  timeVar,
  intakeVar,
  model_str = "FPM",
  idVar = NA,
  hessian = FALSE
)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{model_str}{The base model to use--'FPM' for the first principles model and 'Kissileff' for the quadratic model.
Default is 'FPM'.}

\item{idVar}{(optional) A string that is the name of the ID variable in data. Optional: only include if
data has multiple unique IDs that you want to be processed separately. Without input, only 1 set of
fitted values will be returned.}

\item{hessian}{(optional) A logical indicator for using the hessian matrix to calculate standard errors of recovered parameter estimates. Default is FALSE.}
}
\value{
A dataframe with fitted parameter values and all optim outputs
}
\description{
This function provides fitted model parameters for the cumulative
intake curve for each participant/unique ID in a data set. The parameters
are either fit with Kissileff's quadratic model (Kissileff, 1982; Kissileff & Guss, 2001)
or the First Principles Model (Thomas et al., 2017). The models are fit using optim {stats}.
}
\examples{

\dontrun{
}

}
\seealso{
For data details see \code{\link{FPM_Fit}} or \code{\link{Kissileff_Fit}}
}
