% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kissileff_Fit.R
\name{Kissileff_Fit}
\alias{Kissileff_Fit}
\title{Kissileff_Fit: Fits beta coeffiecents for Kissileff's quadratic model}
\usage{
Kissileff_Fit(data, parameters, timeVar, intakeVar)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{CI}{(optional) A logical indicator for whether to return the Hessian matrix derived confidence interval for
for parameters. Default is FALSE}
}
\value{
A list with of values from optim fit
\itemize{
  \item{par}{fitted parameters intercept (par[1]), linear (par[2]), quadratic (par[3])}
  \item{ value}{-2 log-likelihood corresponding to the fitted parameters}
  \item{counts}{the number of itterations to achieve fit}
  \item{convergence}{code for convergence status; see optim}
  \item{message}{a string with additional information, generally NULL}
}
}
\description{
This function fits the beta coefficients for Kissileff's quadratic model
for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001) using
optim {stats}.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the First Principles Model (Thomas et al., 2017),
see \code{\link{FPM_Fit}}.
}
