% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_CIbounds.R
\name{LRT_CIbounds}
\alias{LRT_CIbounds}
\title{LRT_CIbounds: Estimates the confidence bounds using log ratio tests}
\usage{
LRT_CIbounds(
  data,
  parameters,
  min_n2ll,
  paramCI = c("theta", "r"),
  model_str = "FPM",
  timeVar,
  intakeVar,
  bound = "both"
)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c(10, 1, -1)) and FPM_Time needs 2 starting parameters (default is c(10, .10)).}

\item{min_n2ll}{This is a single value or list of the fitted minimum -2 log-likelihood. This will be used as the -2 log-likelihood value for all fitted parameter values entered in paramCI. If multiple values are entered, must enter an equal number of parameter values in paramCI for each labeled parameter. E.g., paramCI = list(r = c(0.10, 0.15)) if enter two -2 log-likelihood values.}

\item{paramCI}{list of strings with the names of the parameters to compute CIs for. Default assumes First Principles Model for both parameters: c('theta'. 'r')}

\item{model_str}{The base model to use--'FPM' for the first principles model and 'Kissileff' for the quadratic model.
Default is 'FPM'.}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{bound}{A string indicating which confidence bound to return: 'upper, 'lower', or 'both'. Default = 'both'}
}
\value{
A list with all information for the specified confidence bounds, model, and parameters
   \item{parFit_names}{A vector of strings with the names of parameters (from entry paramCI)}
   \item{parFit_values}{A vector of parameter values (from entry parameters)}
   \item{parFit_min_n2ll}{The -2 log-likelihood corresponding to the model fit for specificed parameters}
   \item{parCI_upper}{A vector of numbers with the upper confidence bounds for each parameter entered in paramCI}
   \item{parCI_lower}{A vector of numbers with the lower confidence bounds for each parameter entered in paramCI}
   \item{parCI_upper_n2ll}{A vector of numbers with the -2 log-likihood for the model fit when the the upper confidence bounds was found for each parameter entered in paramCI}
   \item{parCI_lower_n2ll}{A vector of numbers with the -2 log-likihood for the model fit when the the lower confidence bounds was found for each parameter entered in paramCI}
   \item{parCI_upper_chisq}{A vector of numbers with the chi-square p-value for the upper confidence bounds for each parameter entered in paramCI}
   \item{parCI_lower_chisq}{A vector of numbers with the chi-square value for the upper confidence bounds for each parameter entered in paramCI}
   \item{parCI_lower_chisq.p}{A vector of numbers with the chi-square value for the lower confidence bounds for each parameter entered in paramCI}
   \item{parCI_upper_chisq.p}{A vector of numbers with the chi-square p-value for the lower confidence bounds for each parameter entered in paramCI}
}
\description{
This function Estimates the confidence bounds using log ratio tests using optim() to identify the bouandries for the upper and lower confidence bounds
}
\examples{

\dontrun{
}

}
