% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_CIbounds.R
\name{LRT_CIbounds}
\alias{LRT_CIbounds}
\title{LRT_CIbounds: Estimates the confidence bounds using log ratio tests}
\usage{
LRT_CIbounds(
  data,
  parameters,
  min_n2ll,
  paramCI = c("theta", "r"),
  fit_fn = FPM_Fit,
  timeVar,
  intakeVar,
  bound = "both"
)
}
\arguments{
\item{data}{A data frame that contains two varliabels:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c(10, 1, -1)) and FPM_Time needs 2 starting parameters (default is c(10, .10)).}

\item{min_n2ll}{This is a single value or list of the fitted minimum -2 log-likelihood. This will be used as the -2 log-likelihood value for all fitted parameter values entered in paramCI. If multiple values are entered, must enter an equal number of parameter values in paramCI for each labeled parameter. E.g., paramCI = list(r = c(0.10, 0.15)) if enter two -2 log-likelihood values.}

\item{fit_fn}{(optional) A string that is the name of the fitting function you want to use: either Kissileff_Fit or
FPM_Fit; default is FPM_Fit.}

\item{timeVar}{A string that is the name of the elapsed time variable in data}

\item{intakeVar}{A string that is the name of the cumulative intake variable in data}

\item{bound}{A string indicating which confidence bound to return: 'upper, 'lower', or 'both'. Default = 'both'}

\item{A}{list of strings with the names of the parameters to compute CIs for. Default assumes First Principles Model for both parameters: c('theta'. 'r')}
}
\value{
NEED TO EDIT
}
\description{
This function Estimates the confidence bounds using log ratio tests using optim() to identify the bouandries for the upper and lower confidence bounds
}
\examples{

\dontrun{
}

}
\seealso{
This uses the original fit functions for the models
(Kissileff, 1982; Kissileff & Guss, 2001; Thomas et al., 2017), see \code{\link{Kissileff_Fit}}, see \code{\link{FPM_Fit}}.
}
