% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamRecovery.R
\name{ParamRecovery}
\alias{ParamRecovery}
\title{IntakeModelParams: Fits model parameters for cumulative intake curves}
\usage{
ParamRecovery(
  nBites,
  Emax,
  parameters,
  time_fn,
  fit_fn,
  simVar,
  simValue,
  nSims,
  bitesize_sd,
  keepBites,
  CI
)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c()) and FPM_Time needs 2 starting parameters (default is c(20, .20)). If enter an original time estimation function you MUST enter the required parameters for that function.}

\item{time_fn}{(suggested) A string that is the name of the time function you want to use to estimate bite timing: either Kissileff_Time or FPM_Time; default is FPM_Time. Can also enter an original fucntion to estimate time, just be sure to include nBites, Emax, and parameters as input arguments to your function and to specify the required parameters.}

\item{fit_fn}{(optional) A string that is the name of the fitting function you want to use: either Kissileff_Fit or
FPM_Fit; default is FPM_Fit. Can also enter an original fucntion to estimate time, just be sure to include nBites,
Emax, and parameters as input arguments to your function and to specify the required parameters.}

\item{simVar}{This indicates what manipuation will be used in the simulation:
-bitesSampled: will randomly select the number of bites from a Gaussian distribution with
mean equal to the number of bites specified (default)
-nbiteError: will randomly alter the number of bites by the amount specified
-mealdurError: will alter the length of the meal in minutes by the amount specified
-EmaxError: will alter the total intake (Emax) but the amount speified
-timpointsError: run simulations where a certain number of timepoints are off in their timing}

\item{simValue}{This is the value used to altered the data as specified in simVar (use negative
valuse to simulate missed bites or decreased meal duration or total intake):
   -bitesSampled: enter standard deviation for the distribution number of bites will be sampled from
   -nbiteError: amount to adjust bites
   -mealdurError: an object of length two specifying meal duration in minutes and number of minutes
   to adjust meal by, e.g., c(30, 3)
   -EmaxError: number of grams to adjust total intake by
   -timpointsError: an object of length two specifying number of timepoints that will have altered
    timing and the amount to alter the timpoint in minutes, e.g., c(3, 0.25)}

\item{nSims}{The number of iterations to use. Default is 500.}

\item{keepBites}{This is a logical indicator for wether to return the simulated bite dataset with
elapsed timem and cumulative intake for each bite across all simulated intake curves. Default is FALSE.}

\item{CI}{(optional) A logical indicator for whether to return the Hessian matrix derived confidence interval for
for parameters. Default is FALSE}
}
\value{
NEED TO EDIT
}
\description{
This function provides runs Monte Carlo simulations for *n* itterations
by simulating the cumulative intake curve usings average bites size and then fitting
the model parameters for each curve. The simulation can be run while altering number of bites or
by adding error to the data by altering number of bites, meal duration, total intake, or the
timing of the bites. There is the option of altering bites size by a given standard deviation to
reflect within meal variation in bite size. The parameters will be fit using either Kissileff's
 quadratic model (Kissileff, 1982; Kissileff & Guss, 2001) or the First Principles Model
 (Thomas et al., 2017).
}
\examples{

\dontrun{
}

}
\seealso{
This function relies on \code{\link{simBites}} and \code{\link{IntakeModelParams}}.
}
