% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quad_Fit.R
\name{Quad_Fit}
\alias{Quad_Fit}
\title{Quad_Fit: Fits beta coefficients for the Quadratic model}
\usage{
Quad_Fit(data, parameters, timeVar, intakeVar)
}
\arguments{
\item{data}{A data frame that contains:
1) elapsed times for each bite/cumulative intake; 2) cumulative intake corresponding to each elapsed time}

\item{parameters}{A set of numeric beta coefficients for the quadratic model in the format: c(intercept, linear, quadrtic).}

\item{timeVar}{(optional) Use if name of bite timing variable does not include 'time' or multiple variable names in data include the string 'time'}

\item{intakeVar}{(optional) Use if name of bite timing variable does not include 'cumulative' or multiple variable names in data include the string 'cumulative'}
}
\value{
A list with of values from optim fit
\itemize{
  \item{par}{fitted parameters intercept (par[1]), linear (par[2]), quadratic (par[3])}
  \item{value}{-2 log-likelihood corresponding to the fitted parameters}
  \item{counts}{the number of itterations to achieve fit}
  \item{convergence}{code for convergence status; see optim}
  \item{message}{a string with additional information, generally NULL}
}
}
\description{
This function fits the beta coefficients for the Quadratic model
for the cumulative intake curves from Kissileff (1982) and Kissileff & Guss (2001) using
optim {stats}.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Logistic Ordinary Differential Equation
(LODE) Model (Thomas et al., 2017),
see \code{\link{LODE_Fit}}.
}
