% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSEcalc.R
\name{RMSEcalc}
\alias{RMSEcalc}
\title{RMSEcalc: Wrapper of model fit functions that calculates the root mean squared error}
\usage{
RMSEcalc(
  data,
  parameters,
  timeVar,
  intakeVar,
  model_str = "FPM",
  error_outcome = "timing",
  Emax
)
}
\arguments{
\item{data}{The dataset with 'True' bite timing and cumulative intake.}

\item{parameters}{A set of numeric parameters for the First Principles Model in the format: c(theta, r)}

\item{timeVar}{The variable name for the 'True' bite timing in dataset}

\item{intakeVar}{The variable name for the 'True' cumulative intake in dataset}

\item{error_outcome}{Which variable to use to calculate error - 'Time' will use bite timing and 'intake' will use cumulative intake. Default is 'timing'.}

\item{Emax}{The total cumulative intake at the end of the meal.}

\item{par}{Fitted parameters for which error is being tested}
}
\value{
NEED TO EDIT
}
\description{
This function wrpas model fit functions called through optim {stats} to calcualte the likelihood ratio test. This is a function that, when called by optim {stats} will identify a confidence bound of a parameter by optim's minimization of the calculated LRT value.
}
\examples{

\dontrun{
}

}
\seealso{
To get fit your intake data using the Kisslieff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001), see \code{\link{Kissileff_Fit}}.
}
