% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBites.R
\name{simBites}
\alias{simBites}
\title{simBites: Simulates cumulative intake for bites and bite timing}
\usage{
simBites(
  nBites,
  Emax,
  parameters,
  model_str = "FPM",
  id = NA,
  procNoise = TRUE,
  biteSize_sd = NA,
  maxDur = NA
)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c(10, 1, -1)) and FPM_Time needs 2 starting parameters (default is c(10, .10)).}

\item{model_str}{The base model to use--'FPM' for the first principles model and 'Kissileff' for the quadratic model.
Default is 'FPM'.}

\item{procNoise}{(optional) A logical indicator for adding random process noise to the bite data by jittering bite size with bite timing estrimated from jittered bite sizes. This uses the default jitter amount (smallest distance/5). Default value is TRUE.}

\item{biteSize_sd}{(optional) This allows you to enter the standard deviation of individuals bites sizes and will replace the default procNoise routine (jittered bite sizes). Bite sizes will be randomly chosen from a normal distribution truncated at min = 0 with mean = Emax/nBites and standard deviation equal to the entered value. procNoise must be set to TRUE, otherwise this argument will be ignored.}

\item{maxDur}{(optional) A numeric value; the maximum meal duration. Used for simulation purposes if using the First Principles model - will check to see if meal duration extends beyond entered value and sample bites based on the Emax possible the given meal duration. Do not need for the Kissileff model.}

\item{idVar}{(optional) A string or numeric value for ID to be added to the simulated bite data.}
}
\value{
A bite dataset with bite timing, bite size, and cumulative intake for each bite
}
\description{
This function generates a bite data set that incluneds cumulative intake for *n* bites
and the estimated bite timing. The simulation calculates cumulative intake using
average bite size so the bite size is the same across the meal. There is the option of altering
bites size by a given standard deviation to reflect within meal variation in bite size.
If sd_bitesize is specified, the bites size will randomly vary across the meal using a Gussian
ditribution with mean = average bite size and SD = sd_bitesize. Once the cumulative intake
is generated, the timing of each bite will be estimated from either the Kissileff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001) or the First Principles Model (Thomas et al., 2017)
cumulative intake models.
}
\examples{

\dontrun{
}


}
