% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBites.R
\name{simBites}
\alias{simBites}
\title{simBites: Simulates cumulative intake for bites and bite timing}
\usage{
simBites(nBites, Emax, bitesize_sd, parameters, time_fn, idVar)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c()) and FPM_Time needs 2 starting parameters (default is c(20, .20)). If enter an original time estimation function you MUST enter the required parameters for that function.}

\item{time_fn}{(suggested) A string that is the name of the time function you want to use to estimate bite timing: either Kissileff_Time or FPM_Time; default is FPM_Time. Can also enter an original fucntion to estimate time, just be sure to include nBites, Emax, and parameters as input arguments to your function and to specify the required parameters.}

\item{idVar}{(optional) A string or numeric value for ID to be added to the simulated bite data.}

\item{sd_bitesize}{(optional) A numeric value indicating the standard deviation for bite size within
a meal.}
}
\value{
NEED TO EDIT
}
\description{
This function generates a bite data set that incluneds cumulative intake for *n* bites
and the estimated bite timing. The simulation calculates cumulative intake using
average bite size so the bite size is the same across the meal. If sd_bitesize is specified,
the bites size will randomly vary across the meal using a Gussian ditribution with mean = average
bite size and SD = sd_bitesize. Once the cumulative intake is generate, the timing of each bite will
be estimated from either the Kissileff's quadratic model (Kissileff, 1982; Kissileff & Guss, 2001)
or the First Principles Model (Thomas et al., 2017) cumulative intake models.
}
\section{}{

}

\examples{

\dontrun{
}

}
\seealso{
See also \code{\link{simCumulativeIntake}} to simualte a cumulative intake curve accross participatns.

Other aggregate functions: 
\code{\link{FPM_Fit}()},
\code{\link{FPM_Intake}()},
\code{\link{FPM_Time}()},
\code{\link{Kissileff_Fit}()},
\code{\link{Kissileff_Intake}()},
\code{\link{Kissileff_Time}()}
}
\concept{aggregate functions}
