% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBites.R
\name{simBites}
\alias{simBites}
\title{simBites: Simulates cumulative intake for bites and bite timing}
\usage{
simBites(
  nBites,
  Emax,
  parameters,
  time_fn = FPM_Time,
  idVar = NA,
  procNoise = FALSE,
  bitesize_sd = NA,
  maxDur = NA,
  intake_fn = FPM_Intake
)
}
\arguments{
\item{nBites}{A numeric value that represents total number of bites in a meal.}

\item{Emax}{A numeric value that represents total cumulative intake.}

\item{parameters}{(suggested) A set of numeric parameters for the bite time estimation function: Kissileff_Time needs 3 starting parameters (default is c(10, 1, -1)) and FPM_Time needs 2 starting parameters (default is c(10, .10)). If enter an original time estimation function you MUST enter the required parameters for that function.}

\item{time_fn}{(suggested) A string that is the name of the time function you want to use to estimate bite timing: either Kissileff_Time or FPM_Time; default is FPM_Time. Can also enter an original fucntion to estimate time, just be sure to include nBites, Emax, and parameters as input arguments to your function and to specify the required parameters.}

\item{idVar}{(optional) A string or numeric value for ID to be added to the simulated bite data.}

\item{procNoise}{(optional) A logical indicator for adding random process noise to the bite data by jittering bite size (and thus estimated timing). Used primarily for stimulations. Default value is FALSE.}

\item{bitesize_sd}{(optional) A numeric value; the standard deviate in an inidvidual's bite size over the course of the meal. This will be the standard deviation of the truncated normal distribution with mean Emax that bite sizes are chosesn from. Need to use procNoise = TRUE to use.}

\item{maxDur}{(optional) A numeric value; the maximum meal duration. Used for simulation purposes. Will check to see if meal duration extends beyoned entered value and sample bites based on the Emax possible the given meal duration. Only used if the quantitative model uses an Emax so is not relevant for the Kissileff model.}

\item{intake_fn}{(optional) Name of function to calculate intake at Emax. Only needed if enter maxDur and you are using your own function. If using Kissileff_Time or FPM_Time, it will automatically use the correct intake function.}
}
\value{
NEED TO EDIT
}
\description{
This function generates a bite data set that incluneds cumulative intake for *n* bites
and the estimated bite timing. The simulation calculates cumulative intake using
average bite size so the bite size is the same across the meal. There is the option of altering
bites size by a given standard deviation to reflect within meal variation in bite size.
If sd_bitesize is specified, the bites size will randomly vary across the meal using a Gussian
ditribution with mean = average bite size and SD = sd_bitesize. Once the cumulative intake
is generated, the timing of each bite will be estimated from either the Kissileff's quadratic model
(Kissileff, 1982; Kissileff & Guss, 2001) or the First Principles Model (Thomas et al., 2017)
cumulative intake models.
}
\examples{

\dontrun{
}

}
\seealso{
See also \code{\link{simCumulativeIntake}} to simualte a cumulative intake curve accross participatns.
}
