% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simParamRecovery.R
\name{simParamRecovery}
\alias{simParamRecovery}
\title{simParamRecovery: This function recovers cumulative intake model parameters for simulated bite data}
\usage{
simParamRecovery(
  data,
  nBites,
  Emax,
  parameters,
  id,
  model_str = "LODE",
  procNoise = TRUE,
  measureNoise = FALSE,
  pNoiseSD = NA,
  mNoise_TimeSD = NA,
  mNoise_IntakeCat = "mean",
  keepBites = FALSE,
  conf = 95,
  error_method,
  error_measure,
  adjustNA = "interpolate",
  distinct = FALSE,
  cutoff
)
}
\arguments{
\item{data}{A data.frame containing nBites, Emax, the parameter values, and id (if id is desired). Note: the naming of the variables must match exactly if using this method. Variable names for paremters must be either 'theta' and 'r' or 'int', 'linear', and 'quad. If no data is enter, must enter each variable individually.}

\item{nBites}{If not using data.frame, a numeric value or a vector of values for total number of bites in a meal.}

\item{Emax}{If not using data.frame, a numeric value or a vector of values for total cumulative intake.}

\item{parameters}{If not using data.frame, a set of numeric parameters or vector of parameter values: the Quadratic Model needs an intercept, linear slope, and quadratic slope entered in that order (e.g., c(10, 1, -1)) and Logistic Ordinary Differential Equation (LODE) model needs theta and r entered in that order (e.g., c(10, .10)).}

\item{id}{(optional)  If no data entered, a numeric value or a vector of values for simulation ids. There should be one id for each set of values entered in the same order}

\item{model_str}{(optional) Only needed if simulating data. 'LODE' for the Logistic Ordinary Differential Equation model and 'Quad' for the Quadratic model. Default is 'LODE'}

\item{procNoise}{(optional) For simulation only (i.e., timeDat not specified). A logical indicator for adding random process noise to the bite data by jittering bite size with bite timing calculated from jittered bite sizes. This uses the default jitter amount (smallest distance/5). Default value when timeDat is TRUE}

\item{measureNoise}{(optional) A logical indicator for measurement noise after the estimation of bite timings. Default value is FALSE.}

\item{pNoiseSD}{(optional) Only use if want to replace default approach (jitter). Standard deviation of individuals to use to generate a Gaussian distribution. Mean will be set depending on entered type: intake - mean bite size; time - mean inter-bite-interval. Distribution will be truncated at 0 if needed.}

\item{mNoise_TimeSD}{(optional) Use only if want to replace default of jitter appraoch. Measurement noise added to each timepoint will be chosen from a Gaussian distribution with mean = 0 and entered standard deviation entered. measureNoise must be set to to 'BiteTiming' or 'Both' otherwise this argument will be ignored. Note: the normal distribution will be truncated at at each timepoint so that the time for timepoint t is not less than timepoint t-1.}

\item{mNoise_IntakeCat}{(option) Use only if want to replace default of using average bite size. Cut points must equal n - 1 categories (e.g., if want three small, medium, and large bitecategories, enter the small-medium and medium-large  cut points). Bite sizes within each category will be set to the average bite size for that category. measureNoise must be set to to 'BiteSize' or 'Both' otherwise this argument will be ignored. Note: Cut points will be left/lower inclusive but exclude upper boundary}

\item{keepBites}{(optional) This is a logical indicator for whether to return the simulated bite dataset with
elapsed time and cumulative intake for each bite across all simulated intake curves. The returned cumulative
intake data will use the same bite timing as in the initial data but will estimate intake for each bite based on
the recovered model parameters. Default is FALSE.}

\item{conf}{(optional) Level of confidence for calculation of confidence intervals around the fitted parameter estimates. Default is 95 for 95\% CI. If no confidence intervals are desired, set conf = NA}

\item{error_method}{(optional) A string indicating which error metrics to compute - 'rmse' for root mean squared error, 'R2' for pseudo-R2, and 'both' for both. If not specified, no error is not computed.}

\item{error_measure}{Which variable to use to calculate error - 'timing' will use bite timing, 'intake' will use cumulative intake, 'both' will return both. Default is 'intake'.}

\item{adjustNA}{The method used to adjust predicted values that are not real integers, resulting in NA values. 'interpolate' will average the true data values from adjacent bites while 'minmax' will use the minimum or maximum true data values for bites at the start and end of the meal, respectively. Default is 'interpolate'.}

\item{distinct}{Indicate if want to calculate parameter estimate distinctness - TRUE or FALSE. Default is FALSE.}
}
\value{
Either 1 or 2 datasets. It will always return a dataset with recovered parameters but will only return a list with bite data sets for each simulation if keepBites = TRUE
}
\description{
This function simulates the cumulative intake curve using average bites size and then fitting the model parameters for each curve. Process noise can be used rather than average bite size, if wanted. Additionally, measurement error can be added after the estimation of bite timing (from bite size) by reverting to average bite size or categorizing bite sizes and jittering the bite timing. The distinction between processes and measurement noise is that process noise is added before the calculation of bite timing while measurement noise is added after and there is no adjustment to fit the model. The parameters will be fit using either the Quadratic model (Kissileff, 1982; Kissileff & Guss, 2001) or the Logistic Ordinary Differential Equation (LODE) Model (Thomas et al., 2017), total intake (Emax), and number of bites.
}
\examples{

\dontrun{
}

}
\seealso{
This function relies on \code{\link{n2LL_LODE}} and \code{\link{n2LL_Quad}}.
}
